+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.1.0.dev0 (unreleased)
+## 0.1.0 (2017-10-14)
 
 ### Added